# docs/일로ON_openapi.yaml
openapi: 3.0.1
info:
 title: 일로ON API
 description: 업무관리 + AI 회의록 서비스 REST API
 version: v0.1
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
 /api/meetings/{meetingId}:
  get:
   tags:
   - Meeting
   summary: 회의 상세
   operationId: get_3
   parameters:
   - name: meetingId
     in: path
     required: true
     schema:
      type: integer
      format: int64
   responses:
    "200":
     description: OK
     content:
      '*/*':
       schema:
        $ref: "#/components/schemas/MeetingDetailResponse"
  put:
   tags:
   - Meeting
   summary: 회의 수정
   operationId: update_3
   parameters:
   - name: meetingId
     in: path
     required: true
     schema:
      type: integer
      format: int64
   requestBody:
    content:
     application/json:
      schema:
       $ref: "#/components/schemas/MeetingUpdateRequest"
    required: true
   responses:
    "200":
     description: OK
     content:
      '*/*':
       schema:
        $ref: "#/components/schemas/MeetingDetailResponse"
  delete:
   tags:
   - Meeting
   summary: 회의 삭제
   operationId: delete_3
   parameters:
   - name: meetingId
     in: path
     required: true
     schema:
      type: integer
      format: int64
   responses:
    "204":
     description: No Content
...
 /api/meetings/{meetingId}/analyze:
  post:
   tags:
   - AI
   summary: "회의 분석 — audio 파일이 있으면 STT 후 분석, 없으면 회의 내용 분석"
   operationId: analyze
   parameters:
   - name: meetingId
     in: path
     required: true
     schema:
      type: integer
      format: int64
   requestBody:
    content:
     multipart/form-data:
      schema:
       type: object
       properties:
        audio:
         type: string
         format: binary
     application/json:
      schema:
       type: object
       properties:
        audio:
         type: string
         format: binary
     application/octet-stream:
      schema:
       type: object
       properties:
        audio:
         type: string
         format: binary
   responses:
    "200":
     description: OK
     content:
      '*/*':
       schema:
        $ref: "#/components/schemas/BriefingResponse"
...
 /api/meetings/{meetingId}/summary:
  get:
   tags:
   - AI
   summary: 회의 브리핑 결과 조회
   operationId: summary
   parameters:
   - name: meetingId
     in: path
     required: true
     schema:
      type: integer
      format: int64
   responses:
    "200":
     description: OK
     content:
      '*/*':
       schema:
        $ref: "#/components/schemas/BriefingResponse"
...
components:
 schemas:
...
  MeetingUpdateRequest:
   type: object
   properties:
    title:
     maxLength: 200
     minLength: 0
     type: string
    content:
     type: string
    meetingAt:
     type: string
     format: date-time
    attendeeIds:
     type: array
     items:
      type: integer
      format: int64
  Attendee:
   type: object
   properties:
    userId:
     type: integer
     format: int64
    name:
     type: string
    email:
     type: string
  MeetingDetailResponse:
   type: object
   properties:
    meetingId:
     type: integer
     format: int64
    projectId:
     type: integer
     format: int64
    projectName:
     type: string
    title:
     type: string
    content:
     type: string
    meetingAt:
     type: string
     format: date-time
    attendees:
     type: array
     items:
      $ref: "#/components/schemas/Attendee"
    hasSummary:
     type: boolean
    taskCount:
     type: integer
     format: int64
    createdAt:
     type: string
     format: date-time
...
  ActionPointResponse:
   type: object
   properties:
    title:
     type: string
    assigneeHint:
     type: string
    dueDate:
     type: string
     format: date-time
    priority:
     type: string
  BriefingResponse:
   type: object
   properties:
    meetingId:
     type: integer
     format: int64
    overview:
     type: string
    highlights:
     type: array
     items:
      type: string
    decisions:
     type: array
     items:
      type: string
    actionPoints:
     type: array
     items:
      $ref: "#/components/schemas/ActionPointResponse"
    source:
     type: string
    provider:
     type: string
    analyzedAt:
     type: string
     format: date-time
...
 securitySchemes:
  bearerAuth:
   type: http
   scheme: bearer
   bearerFormat: JWT